+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Fixed
+- Added some new externs definitions to Cognito externs file (backported from one customer project)
 
 ## [0.3.8] - 2020-05-01
 ### Changed