+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Changed
+- Use newer versions of dependencies
+
 ### Added
 - Added `+hydrogen/scheduling.twarc` profile, for job scheduling using `magnet/scheduling.twarc` Duct library.