+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+## [0.4.0] - 2020-05-03
 ### Changed
 - Use newer versions of dependencies
 
@@ -128,8 +130,9 @@ were moved to an index of suggested next steps in the relevant `.md` file.
 Sometimes it was `re-frame` and sometimes it was `rf`.
 This version unifies it to `rf`.
  
-[UNRELEASED]:  https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.3.9...HEAD
-[0.3.9]: https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.3.7...v0.3.9
+[UNRELEASED]:  https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.3.9...v0.4.0
+[0.3.9]: https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.3.8...v0.3.9
 [0.3.8]: https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.3.7...v0.3.8
 [0.3.7]: https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.3.6...v0.3.7
 [0.3.6]: https://github.com/magnetcoop/hydrogen.duct-template/compare/v0.3.5...v0.3.6